# gh-119396: Optimize unicode_repr()