@vstinner 2024-05-27
Benchmark:

```py
import pyperf
runner = pyperf.Runner()
squote = "'"
dquote = '"'
nl = '\n'
runner.bench_func("repr('abc')", repr, 'abc')
runner.bench_func("repr('a' * 100)", repr, 'a' * 100)
runner.bench_func("repr(('a' + squote) * 100)", repr, ('a' + squote) * 100)
runner.bench_func("repr(('a' + nl) * 100)", repr, ('a' + nl) * 100)
runner.bench_func("repr(dquote + ('a' + squote) * 100)", repr, dquote + ('a' + squote) * 100)
```