@vstinner 2024-05-27
This is a first step. The second step will be to avoid a temporary string in `PyUnicode_FromFormat("%R", str_obj)`.