@vstinner 2024-05-28
> This is a first step. The second step will be to avoid a temporary string in PyUnicode_FromFormat("%R", str_obj).

I implemented the second step locally. Sadly, it's slower! Not faster. IMO the first step (making the code faster) is still worth it :-)